Slight refactor and essential UT (#1308)
* slight refactor and essential UT

* fmt update
